Update naming and finalize
